+++ Servant_Remote_Station/SubGHz_Phy/App/subghz_phy_app.h
@@ -41,7 +41,7 @@ struct __attribute__((packed)) dataPacket {
   int32_t temperature;
   uint32_t pressure;
   uint32_t humidity;
-  uint16_t seqNum;
+  uint8_t seqNum;
   uint8_t nodeID;
 };
 
@@ -79,8 +79,8 @@ extern bool packetACKED;
 
 /* Exported functions prototypes ---------------------------------------------*/
 /**
-  * @brief  Init Subghz Application
-  */
+ * @brief  Init Subghz Application
+ */
 void SubghzApp_Init(void);
 
 /* USER CODE BEGIN EFP */
